dx-compile: load log-meta transform from a vite-free subpath
`dx-compile` lazily imports `@dxos/vite-plugin-log` to inject `@dxos/log`
call-site meta. It imported the package barrel, which also exports the Vite
plugin and therefore evaluates `vite` — inside an esbuild build that is itself
running under a Vite module runner that fails, producing spurious build noise:

  dx-compile: failed to load `@dxos/vite-plugin-log`; continuing without
  log-meta injection. TypeError: First argument must be an Error object
  ... ReferenceError: Cannot access 'transformLogMeta' before initialization.

- Add a `./transform` export to `@dxos/vite-plugin-log` (no `vite` import on
  that module's graph) and import it instead of the barrel.
- Extract the duplicated loader out of the SWC and Babel/Solid plugins into
  `load-log-meta-transform.ts`.
- Treat `ERR_PACKAGE_PATH_NOT_EXPORTED` as the expected self-bootstrapping
  case, and format the warning's error as a string so a non-Error rejection
  cannot mask the real cause.
